Reduce background mesh size for detection on coadds
Until now, only the threshold type and value for detections were
defined in this config.

With this commit, I've reduced the background mesh size for detection
on coadds by setting binSize=128 pixels, consistent with
obs_subaru. This adjustment reduces the load on the deblender and
improves detection and photometry quality for small, nearly unresolved
sources. For those focusing on large sources, the background can be
restored for specialized measurements.