Switch to CalibrateImageTask
Until now, we used the "characterizeImage", "calibrate", and
"writeSourceTable" tasks in step 1 of pipeline, but now the
"calibrateImage" task has replaced all of them.

"CalibrateImageTask" performs “single frame processing” on one single
visit post-Instrument Signature Removal single detector exposure
(postISRCCD). This involves repairing cosmic rays and defects,
detecting and measuring sources on the exposure to make an initial
estimation of the point spread function (PSF), re-doing detection and
measurement with that fitted PSF to compute the astrometric and
photometric calibrations, and computing summary statistics of the
exposure and measured catalog.

With this commit, I updated the DRP YAML file and added the
"calibrateImage" configuration files to the configs.  This task also
requires the "computeExposureSummaryStats.py" config to set parameters
like "fiducialPsfSigma", "fiducialSkyBackground", and
"fiducialZeroPoint".

For now, I've commented these parameters out, and we can add them
later once they are determined.